requirements.txt
Pillow==9.3.0
rembg==2.0.31
streamlit==1.24.1
